requirements.txt
fastapi
fire
gitpython
requests
uvicorn
